Statistics for game units in Age of Empires II: Definitive Edition

A dataset with the numerical values that determine the behavior and performance of selected military units available in AoE2:DE (July 2020 Game Update).

Usage

AOEunits

Format

A data frame with 128 observations of 19 variables:

unit

Unit name

building

Building in which each unit is trained

type

Unit class

age

Age at which the unit becomes trainable

cost_wood

Unit cost in Wood

cost_food

Unit cost in Food

cost_gold

Unit cost in Gold

build_time

Training time in seconds

rate_of_fire

Attack speed

attack_delay

Retasking time

movement_speed

Travel speed on land

line_of_sight

Vision over the surrounding area

hit_points

Unit health

min_range

Minimum attacking range for ranged units

range

Maximum attacking range for ranged units

damage

Damage inflicted per attack

accuracy

Chance that an attack will be on target

melee_armor

Armor against melee attacks

pierce_armor

Armor against projectiles

Source

Age of Empires II. Copyright Microsoft Corporation. This dataset was created under Microsoft's Usage rules for Digital Goods using assets from Age of Empires II, and it is not endorsed by or affiliated with Microsoft. All information shown is an interpretation of data collected in-game with no guarantee on the accuracy of any of the data presented.