Annotate meaningful formatting for all cells

Turns cell formatting into annotations for all values across all variables.

annotate_mf_all(xlfilepath)

Arguments

xlfilepath

Path to a single-sheet spreadsheet file (xls or xlsx).

Value

A tibble with meaningful formatting embedded as text for all rows and columns.

Details

Seven popular approaches for meaningful formatting (bold, colored text, italic, strikethrough, underline, double underline, and cell highlighting) are supported in this function. sheets, skip, and range arguments for spreadsheet input are not supported. The hex8 code of the fill color used for text color and cell highlighting is also appended in the output. Ensure the data in the spreadsheet are rectangular before running; this includes blank but formatted cells beyond the data rectangle.

Examples

example_spreadsheet <- system.file("extdata/boutiques.xlsx", package = "unheadr")
annotate_mf_all(example_spreadsheet)
#> # A tibble: 7 × 6
#>   `Store Location`                  Q1_2012      Q1_2013 Q1_2014 Q1_2015 Q1_2016
#>   <chr>                             <chr>        <chr>   <chr>   <chr>   <chr>  
#> 1 (highlight-FFADC5E7) London       (highlight-… (bold,… (highl… (highl… (bold,…
#> 2 (highlight-FFADC5E7) Paris        (highlight-… (highl… (highl… (bold,… (highl…
#> 3 (highlight-FFF37B70) Atlanta      (bold, high… (highl… (highl… (highl… (highl…
#> 4 (highlight-FFADC5E7) Madrid       (highlight-… (bold,… (bold,… (bold,… (highl…
#> 5 (highlight-FFADC5E7) Rome         (highlight-… (highl… (highl… (highl… (highl…
#> 6 (highlight-FFF37B70) Mexico City  (bold, high… (highl… (bold,… (highl… (bold,…
#> 7 (highlight-FFADC5E7, italic) Rome (highlight-… (highl… (highl… (highl… (highl…